+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file, following t
 - chainname based color scheme
 - BondHash class to quickly get atoms connected to an atom
 - XmlParser parameter to use the browser's DOMParser
+- label attachment (top, middle, bottom; left, center, right)
 
 ### Changed
 - ResidueindexColorMaker colorscale domain on a per chain basis